+++ ivy/functional/backends/jax/extensions/manipulation.py
@@ -1,0 +1,33 @@
+from typing import Optional, Union, Sequence
+from ivy.functional.backends.jax import JaxArray
+import jax.numpy as jnp
+
+
+def moveaxis(
+    a: JaxArray,
+    source: Union[int, Sequence[int]],
+    destination: Union[int, Sequence[int]],
+    /,
+    *,
+    out: Optional[JaxArray] = None,
+) -> JaxArray:
+    return jnp.moveaxis(a, source, destination)
+
+
+def heaviside(
+    x1: JaxArray,
+    x2: JaxArray,
+    /,
+    *,
+    out: Optional[JaxArray] = None,
+) -> JaxArray:
+    return jnp.heaviside(x1, x2)
+
+
+def flipud(
+    m: JaxArray,
+    /,
+    *,
+    out: Optional[JaxArray] = None,
+) -> JaxArray:
+    return jnp.flipud(m)
